Refactor to prioritize JNA native calls and clean code.
Replaced fallback ProcessBuilder calls with JNA native calls across key methods like `setHostname`, `umount`, `reboot`, and others. Removed commented-out code and improved error handling with clear logging messages. Enhanced clarity by simplifying variable definitions and maintaining consistency.

+++ starship-sdk/src/main/groovy/org/starship/jna/CLib.groovy
@@ -23,7 +23,7 @@ interface CLib extends Library {
     int S_IWUSR = 0x0080 // Write permission, owner
 
     // Reboot command
-    int LINUX_REBOOT_CMD_HALT = 0xCDEF0123
+    int LINUX_REBOOT_CMD_HALT = (int) 0xCDEF0123
 
     // Load the native library
     CLib INSTANCE = Native.load("c", CLib.class)
